Server: KiteServer complete update for es6
